fix(ui): resolve glossary lazy split imports
